// src/app/[locale]/applicant/resume/(edit)/personal/personal.d.ts
export type PersonalDTO = {
  "category": string | undefined,
  "first_name": string,
  "last_name": string | undefined,
  "father_name": string,
  "mother_name": string | undefined,
  "dob": string,
  "gender": string,
  "religion": string,
  "marital_status": string,
  "nationality": string,
  "nid": string | undefined,
  "primary_mobile": string,
  "secondary_mobile": string | undefined,
  "primary_email": string,
  "alternate_email": string | undefined,
  "height": number | undefined,
  "weight": number | undefined,
  "present_address": string,
  "permanent_address": string | undefined,
  "is_same_address": boolean,
  "present_salary": string | undefined,
  "expected_salary": string | undefined,
  "job_level": string,
  "job_type": string,
  "special_skills": number[] | Array,
  "preferred_functions": number[] | Array,
  "career_objective": string,
  "career_summary": string | undefined,
  "special_qualification": string | undefined,
  "has_disability": number,
  "disability_id": string | undefined
}